# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: UnicornOTP
# during dev, should point to your local machine
host: localhost:10099
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  "/{channel}":
    x-swagger-router-controller: otp_main
    post:
      description: Create an OTP
      operationId: createOTP
      parameters:
        - name: channel
          in: path
          type: string
          description: The channel on which OTP needs to be generated.
        - name: returnOTP
          in: query
          type: boolean
          description: Define if the OTP needs to retruned to the caller
        - name: outbound
          in: query
          type: boolean
          description: Send to a pre-defined outbound URL
        - name: token
          in: query
          type: string
          description: API Token for the channel
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/CreateOTPResponse"
        default:
          description: Failure response
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  CreateOTPResponse:
    properties:
      otp:
        type: string
        description: OTP generated
      context:
        type: string
        description: context id of the generated OTP. You need this context when cross validating the OTP.
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string